+++ cli/src/html/__test__/parser/examples/WrongSignatureNoHtmlTag.figma.ts
@@ -5,6 +5,5 @@ figma.connect('', {
     visible: figma.boolean('Visible'),
     label: figma.string('Label'),
   },
-  // @ts-expect-error
   example: (props) => `<my-component visible=${props.visible}>${props.label}</my-component>`,
 })